:truck: :page_facing_up: Use pdfkit code
Bundle pdfkit code with its license. Ignore it when generating test coverage.
